pyproject.toml
[project]
name = "roma-stt"
version = "0.1.0"
description = "Локальное распознавание речи в текст (Speech-to-Text) для Windows: трей, горячая клавиша, whisper.cpp"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "Pillow",
    "pyperclip",
    "pystray",
    "PyYAML",
    "pywin32",
    "sounddevice",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
gui = [
    "PySide6>=6.6",
]

[project.scripts]
roma-stt-control = "presentation.control_gui.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["application", "domain", "infrastructure", "presentation"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff>=0.9",
]
